requirements.txt
streamlit
pandas
openpyxl
faiss-cpu
google-generativeai
langchain
langchain-community
langchain-google-genai
scikit-learn    